update: update-status-func

+++ src/components/ui/todoItem.tsx
@@ -6,25 +6,31 @@ import { Todo } from "@/types"
 interface TodoItemProps {
   todo: Todo;
   index: number;
-  onDeleteTodo: (index:number)=>void;
+  onUpdateTodo: (index: number, todo: Todo) => void;
+  onDeleteTodo: (index: number) => void;
 }
 
-const TodoItem: React.FC<TodoItemProps> = ({ index, todo, onDeleteTodo }) => {
+const TodoItem: React.FC<TodoItemProps> = ({ index, todo, onUpdateTodo, onDeleteTodo }) => {
   const handleDeleteButtonClick = () => {
-    onDeleteTodo(index);
+    confirm("Do you really remove?") &&
+      onDeleteTodo(index);
+  }
+
+  const handleUpdateButtonClick = () => {
+    onUpdateTodo(index, todo);
   }
 
   return (
     <div className="w-[326px] h-[124px] border border-black px-1 py-2">
       <div className="leading-8">
-        <span className="mr-2">{todo.status == '1' ? FINISH : UNFINISH}</span>
+        <span className="mr-2">{todo.status ? FINISH : UNFINISH}</span>
         <span>{TODO_TITLE}: {todo.title}</span>
       </div>
       <div>
         <span>{TODO_CONTENT}: {todo.content}</span>
       </div>
       <div className="flex float-right gap-2">
-        <UpdateButton />
+        <UpdateButton onUpdateButtonClick={handleUpdateButtonClick} />
         <DeleteButton onDeleteButtonClick={handleDeleteButtonClick} />
       </div>
     </div>
